+++ ControlFactory.cs
@@ -223,6 +223,7 @@ namespace GoatForms
         /// </summary>
         /// <param name="form">The parent form to which the <see cref="ListBox"/> will be added.</param>
         /// <param name="description">The description for the  <see cref="ListBox"/> control.</param>
+        /// <param name="onSelectedIndexChanged">The selection changed <see chref="EventHandler"/> for the  <see cref="ListBox"/> control.</param>
         /// <param name="styledControl">Whether the the <see cref="ListBox"/> control should be styled.</param>
         /// <returns>A handle to the created <see cref="ListBox"/> control.</returns>
         /// <exception cref="ArgumentNullException">Thrown when <paramref name="form"/> is <see langword="null"/>.</exception>
@@ -230,7 +231,7 @@ namespace GoatForms
         /// This method creates a <see cref="ListBox"/> control with the specified properties and adds it to the specified form. 
         /// If the parent control is <see langword="null"/>, an <see cref="ArgumentNullException"/> will be thrown. 
         /// </remarks>
-        public static ListBox AddListBox(BaseForm form, string description = null, bool styledControl = false)
+        public static ListBox AddListBox(BaseForm form, string description = null, EventHandler onSelectedIndexChanged = null, bool styledControl = false)
         {
             if (form == null)
             {
@@ -244,6 +245,11 @@ namespace GoatForms
                 Height = 150,
                 TabStop = true
             };
+
+            if (onSelectedIndexChanged != null)
+            {
+                listBox.SelectedIndexChanged += onSelectedIndexChanged;
+            }
 
             form.AddControl(listBox, styledControl);
 
@@ -277,6 +283,7 @@ namespace GoatForms
         /// </summary>
         /// <param name="form">The parent form to which the <see cref="CheckedListBox"/> will be added.</param>
         /// <param name="description">The description for the  <see cref="CheckedListBox"/> control.</param>
+        /// <param name="onSelectedIndexChanged">The selection changed <see chref="EventHandler"/> for the  <see cref="CheckedListBox"/> control.</param>
         /// <param name="styledControl">Whether the the <see cref="CheckedListBox"/> control should be styled.</param>
         /// <returns>A handle to the created <see cref="CheckedListBox"/> control.</returns>
         /// <exception cref="ArgumentNullException">Thrown when <paramref name="form"/> is <see langword="null"/>.</exception>
@@ -284,7 +291,7 @@ namespace GoatForms
         /// This method creates a <see cref="CheckedListBox"/> control with the specified properties and adds it to the specified form. 
         /// If the parent control is <see langword="null"/>, an <see cref="ArgumentNullException"/> will be thrown. 
         /// </remarks>
-        public static CheckedListBox AddCheckedListBox(BaseForm form, string description = null, bool styledControl = false)
+        public static CheckedListBox AddCheckedListBox(BaseForm form, string description = null, EventHandler onSelectedIndexChanged = null, bool styledControl = false)
         {
             if (form == null)
             {
@@ -298,6 +305,11 @@ namespace GoatForms
                 Height = 150,
                 TabStop = true
             };
+
+            if (onSelectedIndexChanged != null)
+            {
+                checkedListBox.SelectedIndexChanged += onSelectedIndexChanged;
+            }
 
             form.AddControl(checkedListBox, styledControl);